Massive re-write
Wanted to try writing the database without using objects as I realised it was just a map between strings and marathon times.  Wrote interface functions for dealing with this and separated out concerns to simplify the logic.  Realised I needed the map to sort with regards to values and not keys so wrote a template class to implement this (which I will probably separate into its own project and expand on later).