Caught bug where map didn't sort properly
Not sure how I didn't catch this.  Calling std::sort without any special function object/lambda just calls the default comparison for std::pairs and so wasn't doing the job properly.  Decided that adding a sort method would be useful in the interface as I could imagine this being needed and it being a pain for the user to have to specify special comparison functions when it should be clear what sort should do.